requirements.txt
google-adk
fastapi
uvicorn[standard]
requests
python-dotenv